fix bug tim3irq; fix bug brightness;

+++ Drivers/BSP/TIMER/btim.h
@@ -46,7 +46,7 @@ void btim_timx_int_init(uint16_t arr, uint16_t psc);    /* 基本定时器 定时中断初
 void btim_timx_int_stop(void);
 void btim_timx_int_start(void);
 void btim_timx_int2_init(uint16_t arr, uint16_t psc);   // 用作类PWM
-uint8_t LED0_brightness_control(uint16_t brightness);
+uint8_t LED0_brightness_control(uint8_t brightness);
 
 #endif
